Implmented EM using Mixture of Gaussians
It trains on the coordinates of plants in the grid. It calculates the
PDF of x and y coordinates vs probability of a plant existing. We use
the idea that there are clusters of plants and by using a mixture of
gaussian we try to predict where these clusters are. Given the current
position, it will tell you the best location to move next based on where
the probability density function is the highest.